LayoutTests/media/modern-media-controls/volume-slider/volume-slider-value.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/volume-slider.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/volume-slider.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>VolumeSlider</code> value property.");

window.jsTestIsAsync = true;

const slider = new VolumeSlider;
slider.x = 10;
slider.y = 32;
slider.width = 200;

slider.uiDelegate = {

    controlValueWillStartChanging: function()
    {
        debug(`controlValueWillStartChanging`);
    },

    controlValueDidChange: function()
    {
        debug(`controlValueDidChange, value = ${slider.value}`);
    },

    controlValueDidStopChanging: function()
    {
        debug(`controlValueDidStopChanging`);
        finishJSTest();
    }

};

const input = slider.children[1].element;

function dragSlider(fromPx, toPx) {
    const bounds = input.getBoundingClientRect();
    const minX = bounds.left;
    const centerY = bounds.top + bounds.height / 2;

    eventSender.mouseMoveTo(minX + fromPx, centerY);
    eventSender.mouseDown();

    const delta = toPx - fromPx;
    const iterations = Math.abs(delta);
    for (let i = 1; i <= iterations; ++i)
        eventSender.mouseMoveTo(minX + fromPx + i * Math.sign(delta), centerY);

    eventSender.mouseUp();
}

scheduler.frameDidFire = function()
{
    document.body.appendChild(slider.element);

    shouldBe("slider.value", "0");
    debug("");
    debug("Now we'll drag to the slider thumb from 10px to 30px within the slider's bounds");
    dragSlider(10, 30);
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
