LayoutTests/media/modern-media-controls/volume-slider/volume-slider-value-expected.txt

Testing the VolumeSlider value property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.value is 0

Now we'll drag to the slider thumb from 10px to 30px within the slider's bounds
controlValueWillStartChanging
controlValueDidChange, value = 0.029
controlValueDidChange, value = 0.034
controlValueDidChange, value = 0.04
controlValueDidChange, value = 0.045
controlValueDidChange, value = 0.05
controlValueDidChange, value = 0.056
controlValueDidChange, value = 0.061
controlValueDidChange, value = 0.066
controlValueDidChange, value = 0.071
controlValueDidChange, value = 0.077
controlValueDidChange, value = 0.082
controlValueDidChange, value = 0.087
controlValueDidChange, value = 0.093
controlValueDidChange, value = 0.098
controlValueDidChange, value = 0.103
controlValueDidChange, value = 0.108
controlValueDidChange, value = 0.114
controlValueDidChange, value = 0.119
controlValueDidChange, value = 0.124
controlValueDidChange, value = 0.13
controlValueDidStopChanging
PASS successfullyParsed is true

TEST COMPLETE
